Adding intermediary class to set up secret keys for test
